Add files via upload
This project is a continuation of the previous one is it is divided into three parts all of which are connected. 

The exercises are themed. They concern modelling aspects of a pay-to-stay car. The kind of car park in question has a ticket machine at the entrance and a cashier at the exit. A driver, on entering the car park receives a ticket stamped with the arrival time. (The arrival time is also recorded on the magnetic strip on the back.) On exit, the driver gives the ticket to the cashier, the duration of the stay is calculated and from that, how much must be paid.

-Part one involves developing  a Ticket class of object. A Ticket object represents a car park ticket. It has a unique ID and time of issue (24-hour clock). The class has a single constructor that is used to (i) assign the current time as the ticket issue time and to (ii) assign a unique ID. 
-Part two involves developing a Register class of an object. The job of a Register object in the car park simulation is to store all the tickets that have been issued.
When a Ticket is issued, it is stored in the register. When the driver departs, the ticket ID is used to retrieve the Ticket object and calculate the duration of stay.
-Part three involves creating a CarParkSimulator class and using both the Register class and the Ticket class.